requirements.txt
avro-python3==1.8.2
boto3==1.9.83
botocore==1.12.83
Click==7.0
docutils==0.14
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.1.0
python-dateutil==2.7.5
PyYAML==3.13
s3transfer==0.1.13
six==1.12.0
urllib3==1.24.1
